added exported set of valid resource types to the compliance/snapshot_poller src
